Move firm graph building classes to firm.construction
