+++ src/main/java/com/github/firmwehr/gentle/firm/construction/FirmGraphBuilder.java
@@ -1,4 +1,4 @@
-package com.github.firmwehr.gentle.firm;
+package com.github.firmwehr.gentle.firm.construction;
 
 import com.github.firmwehr.gentle.semantic.Namespace;
 import com.github.firmwehr.gentle.semantic.ast.LocalVariableDeclaration;
