+++ src/main/java/com/github/firmwehr/gentle/firm/construction/SlotTable.java
@@ -1,4 +1,4 @@
-package com.github.firmwehr.gentle.firm;
+package com.github.firmwehr.gentle.firm.construction;
 
 import com.github.firmwehr.gentle.parser.ast.Ident;
 import com.github.firmwehr.gentle.semantic.SemanticException;
